# tsdb: Added ChunkQueryable implementations to db; unified MergeSeriesSets and vertical to single struct.