This PR adds support for all block Querier to operate on chunk and allow chunk iterating if needed. To do so, we reused all structs done before for merging queries (needed by fanout). This removes lots of code that was duplicated for same work: Querying, fan-out (e.g to multiple block) and merging results.

## Changes

* Added single SeriesEntry / ChunkEntry for all series implementations.
* Unified all vertical, and non vertical for compact and querying to single
merge series / chunk sets by reusing VerticalSeriesMergeFunc for overlapping algorithm (same logic as before)
* Added block (Base/Chunk/)Querier for block querying. We then use populateAndTomb(Base/Chunk/) to iterate over chunks or samples.
* Refactored endpoint tests and querier tests to include subtests.

Signed-off-by: Bartlomiej Plotka <bwplotka@gmail.com>