@bwplotka 2020-06-26
This should be good to go. The problem is that it ended up be quite big. I think I will split it to two PRs:

* Adding Chunk Iter
* Unifying compact / querier with storage