@bwplotka 2020-06-29
PTAL @brian-brazil @codesome @krasi-georgiev if you have time, also feedback welcome how to split it to multiple PR to make it easier for you. Summary of this PR should be up to date.