@bwplotka 2020-07-10
Thanks @beorn7 @brian-brazil for quick review :heart:  Sorry for lag on this, we are back on track: We added nice benchmarks to Thanos Receive so we have now good benchmarks ready to test against chunk iterator. Next steps for this PR:

1. Fix last unit tests here
2. Run Prombench
3. Address review comments & Merge

In meantime I will play with this change on Thanos and benchmark using https://github.com/thanos-io/thanos/blob/master/pkg/store/multitsdb_test.go#L43

Thanks! @codesome can you take a look as well? :hugs: