@bwplotka 2020-07-10
Tests & benchmarks on Thanos side: https://github.com/thanos-io/thanos/pull/2876