@bwplotka 2020-07-22
Feel free @codesome to propose what I can move to another PR, I tried to trimmed it as much as possible.