@bwplotka 2020-07-22
And thanks for reivewing! :heart: I am finished last tests - they are now mostly some inconsistencies on what empty value is returned `nil` vs `[]string(nil)`