@bwplotka 2020-07-24
Ok tests are passing at this point. This is ready to go, all issues fixed @brian-brazil @codesome @krasi-georgiev @beorn7 



In meantime (in separate PR) I will add some remote read benchmark and run existing ones against Querier and PromQL + querier (: