@roidelapluie 2020-07-24
> /prombench v2.19.2

is there a reason not to compare to master ?