@codesome 2020-07-24
master it should be. 2.20 had some cpu/query perf improvements and the benchmark might show wrong results with 2.19.2.