@roidelapluie 2020-07-24
> There are many unknowns with master and it changes everyday. We want to have stable comparison nothing else. I can point to anything latest, just some fixed commit.. Will use something newer.

Ideally you would rebase your work on top of that then, otherwise you are testing more than your fix.