@bwplotka 2020-07-24
Benchmarks look quite epic as well.

-benchtime=30

Old `93e9c010f3aaf9e683574991080e85a428d554ee` (master)
New: This PR `2605f60545cbf7d1325b8fcc04c3affecb9d206b`
```
name                                          old time/op    new time/op    delta
QuerierSelect/Head/1of1000000-12                 646ms ± 0%     323ms ± 0%  -49.93%
QuerierSelect/Head/10of1000000-12                757ms ± 0%     333ms ± 0%  -55.98%
QuerierSelect/Head/100of1000000-12               694ms ± 0%     339ms ± 0%  -51.13%
QuerierSelect/Head/1000of1000000-12              660ms ± 0%     339ms ± 0%  -48.58%
QuerierSelect/Head/10000of1000000-12             740ms ± 0%     334ms ± 0%  -54.89%
QuerierSelect/Head/100000of1000000-12            777ms ± 0%     339ms ± 0%  -56.35%
QuerierSelect/Head/1000000of1000000-12           987ms ± 0%     631ms ± 0%  -36.07%
QuerierSelect/SortedHead/1of1000000-12           2.05s ± 0%     1.31s ± 0%  -36.37%
QuerierSelect/SortedHead/10of1000000-12          1.82s ± 0%     1.29s ± 0%  -29.00%
QuerierSelect/SortedHead/100of1000000-12         1.92s ± 0%     1.29s ± 0%  -32.54%
QuerierSelect/SortedHead/1000of1000000-12        1.88s ± 0%     1.25s ± 0%  -33.45%
QuerierSelect/SortedHead/10000of1000000-12       1.93s ± 0%     1.27s ± 0%  -34.11%
QuerierSelect/SortedHead/100000of1000000-12      2.05s ± 0%     1.28s ± 0%  -37.26%
QuerierSelect/SortedHead/1000000of1000000-12     2.56s ± 0%     1.50s ± 0%  -41.62%
QuerierSelect/Block/1of1000000-12                1.04s ± 0%     0.66s ± 0%  -37.14%
QuerierSelect/Block/10of1000000-12               1.06s ± 0%     0.69s ± 0%  -34.95%
QuerierSelect/Block/100of1000000-12              1.20s ± 0%     0.60s ± 0%  -49.72%
QuerierSelect/Block/1000of1000000-12             1.11s ± 0%     0.67s ± 0%  -39.61%
QuerierSelect/Block/10000of1000000-12            1.10s ± 0%     0.69s ± 0%  -37.35%
QuerierSelect/Block/100000of1000000-12           1.17s ± 0%     0.78s ± 0%  -33.72%
QuerierSelect/Block/1000000of1000000-12          1.81s ± 0%     1.03s ± 0%  -42.97%

name                                          old alloc/op   new alloc/op   delta
QuerierSelect/Head/1of1000000-12                 176MB ± 0%      48MB ± 0%  -72.73%
QuerierSelect/Head/10of1000000-12                176MB ± 0%      48MB ± 0%  -72.73%
QuerierSelect/Head/100of1000000-12               176MB ± 0%      48MB ± 0%  -72.72%
QuerierSelect/Head/1000of1000000-12              176MB ± 0%      48MB ± 0%  -72.65%
QuerierSelect/Head/10000of1000000-12             177MB ± 0%      50MB ± 0%  -71.96%
QuerierSelect/Head/100000of1000000-12            190MB ± 0%      66MB ± 0%  -65.55%
QuerierSelect/Head/1000000of1000000-12           320MB ± 0%     224MB ± 0%  -30.00%
QuerierSelect/SortedHead/1of1000000-12           229MB ± 0%     101MB ± 0%  -55.85%
QuerierSelect/SortedHead/10of1000000-12          229MB ± 0%     101MB ± 0%  -55.85%
QuerierSelect/SortedHead/100of1000000-12         229MB ± 0%     101MB ± 0%  -55.84%
QuerierSelect/SortedHead/1000of1000000-12        229MB ± 0%     101MB ± 0%  -55.80%
QuerierSelect/SortedHead/10000of1000000-12       231MB ± 0%     103MB ± 0%  -55.36%
QuerierSelect/SortedHead/100000of1000000-12      244MB ± 0%     119MB ± 0%  -51.23%
QuerierSelect/SortedHead/1000000of1000000-12     373MB ± 0%     277MB ± 0%  -25.72%
QuerierSelect/Block/1of1000000-12                227MB ± 0%      99MB ± 0%  -56.34%
QuerierSelect/Block/10of1000000-12               227MB ± 0%      99MB ± 0%  -56.34%
QuerierSelect/Block/100of1000000-12              227MB ± 0%      99MB ± 0%  -56.33%
QuerierSelect/Block/1000of1000000-12             227MB ± 0%      99MB ± 0%  -56.28%
QuerierSelect/Block/10000of1000000-12            229MB ± 0%     101MB ± 0%  -55.80%
QuerierSelect/Block/100000of1000000-12           243MB ± 0%     118MB ± 0%  -51.25%
QuerierSelect/Block/1000000of1000000-12          384MB ± 0%     291MB ± 0%  -24.17%

name                                          old allocs/op  new allocs/op  delta
QuerierSelect/Head/1of1000000-12                 4.00M ± 0%     1.00M ± 0%  -75.00%
QuerierSelect/Head/10of1000000-12                4.00M ± 0%     1.00M ± 0%  -75.00%
QuerierSelect/Head/100of1000000-12               4.00M ± 0%     1.00M ± 0%  -74.99%
QuerierSelect/Head/1000of1000000-12              4.00M ± 0%     1.00M ± 0%  -74.91%
QuerierSelect/Head/10000of1000000-12             4.02M ± 0%     1.04M ± 0%  -74.13%
QuerierSelect/Head/100000of1000000-12            4.20M ± 0%     1.40M ± 0%  -66.67%
QuerierSelect/Head/1000000of1000000-12           6.00M ± 0%     5.00M ± 0%  -16.67%
QuerierSelect/SortedHead/1of1000000-12           4.00M ± 0%     1.00M ± 0%  -75.00%
QuerierSelect/SortedHead/10of1000000-12          4.00M ± 0%     1.00M ± 0%  -75.00%
QuerierSelect/SortedHead/100of1000000-12         4.00M ± 0%     1.00M ± 0%  -74.99%
QuerierSelect/SortedHead/1000of1000000-12        4.00M ± 0%     1.00M ± 0%  -74.91%
QuerierSelect/SortedHead/10000of1000000-12       4.02M ± 0%     1.04M ± 0%  -74.13%
QuerierSelect/SortedHead/100000of1000000-12      4.20M ± 0%     1.40M ± 0%  -66.67%
QuerierSelect/SortedHead/1000000of1000000-12     6.00M ± 0%     5.00M ± 0%  -16.67%
QuerierSelect/Block/1of1000000-12                6.00M ± 0%     3.00M ± 0%  -50.00%
QuerierSelect/Block/10of1000000-12               6.00M ± 0%     3.00M ± 0%  -50.00%
QuerierSelect/Block/100of1000000-12              6.00M ± 0%     3.00M ± 0%  -50.00%
QuerierSelect/Block/1000of1000000-12             6.00M ± 0%     3.00M ± 0%  -49.97%
QuerierSelect/Block/10000of1000000-12            6.04M ± 0%     3.04M ± 0%  -49.67%
QuerierSelect/Block/100000of1000000-12           6.40M ± 0%     3.40M ± 0%  -46.87%
```

And PromQL one for same commits. This is big so collapsible:

<details>
  <summary>Click to expand!</summary>
  
 
```
name                                                                             old time/op    new time/op    delta
RangeQuery/expr=a_one,steps=1-12                                                   19.3µs ± 0%    15.4µs ± 0%  -20.36%
RangeQuery/expr=a_one,steps=10-12                                                  19.9µs ± 0%    16.0µs ± 0%  -19.56%
RangeQuery/expr=a_one,steps=100-12                                                 31.6µs ± 0%    25.3µs ± 0%  -19.74%
RangeQuery/expr=a_one,steps=1000-12                                                 111µs ± 0%      85µs ± 0%  -23.08%
RangeQuery/expr=a_ten,steps=1-12                                                   84.0µs ± 0%    62.5µs ± 0%  -25.61%
RangeQuery/expr=a_ten,steps=10-12                                                  89.1µs ± 0%    80.3µs ± 0%   -9.96%
RangeQuery/expr=a_ten,steps=100-12                                                  186µs ± 0%     169µs ± 0%   -9.33%
RangeQuery/expr=a_ten,steps=1000-12                                                 984µs ± 0%     795µs ± 0%  -19.19%
RangeQuery/expr=a_hundred,steps=1-12                                                778µs ± 0%     584µs ± 0%  -24.88%
RangeQuery/expr=a_hundred,steps=10-12                                               802µs ± 0%     750µs ± 0%   -6.55%
RangeQuery/expr=a_hundred,steps=100-12                                             1.52ms ± 0%    2.30ms ± 0%  +51.45%
RangeQuery/expr=a_hundred,steps=1000-12                                            8.01ms ± 0%    7.95ms ± 0%   -0.70%
RangeQuery/expr=rate(a_one[1m]),steps=1-12                                         21.3µs ± 0%    20.7µs ± 0%   -2.68%
RangeQuery/expr=rate(a_one[1m]),steps=10-12                                        22.6µs ± 0%    20.1µs ± 0%  -11.13%
RangeQuery/expr=rate(a_one[1m]),steps=100-12                                       42.2µs ± 0%    37.2µs ± 0%  -11.88%
RangeQuery/expr=rate(a_one[1m]),steps=1000-12                                       211µs ± 0%     179µs ± 0%  -14.78%
RangeQuery/expr=rate(a_ten[1m]),steps=1-12                                         98.2µs ± 0%    72.7µs ± 0%  -26.02%
RangeQuery/expr=rate(a_ten[1m]),steps=10-12                                         108µs ± 0%      87µs ± 0%  -19.65%
RangeQuery/expr=rate(a_ten[1m]),steps=100-12                                        313µs ± 0%     278µs ± 0%  -11.38%
RangeQuery/expr=rate(a_ten[1m]),steps=1000-12                                      2.08ms ± 0%    1.71ms ± 0%  -17.62%
RangeQuery/expr=rate(a_hundred[1m]),steps=1-12                                      854µs ± 0%     555µs ± 0%  -35.05%
RangeQuery/expr=rate(a_hundred[1m]),steps=10-12                                    1.02ms ± 0%    0.69ms ± 0%  -32.27%
RangeQuery/expr=rate(a_hundred[1m]),steps=100-12                                   3.09ms ± 0%    2.46ms ± 0%  -20.32%
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-12                                  21.8ms ± 0%    16.6ms ± 0%  -23.69%
RangeQuery/expr=rate(a_one[1m]),steps=10000-12                                     2.30ms ± 0%    2.60ms ± 0%  +12.83%
RangeQuery/expr=rate(a_ten[1m]),steps=10000-12                                     22.5ms ± 0%    26.5ms ± 0%  +17.76%
RangeQuery/expr=rate(a_hundred[1m]),steps=10000-12                                  232ms ± 0%     270ms ± 0%  +16.52%
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-12                       1.49ms ± 0%    1.12ms ± 0%  -24.73%
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-12                      1.86ms ± 0%    1.69ms ± 0%   -9.05%
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-12                     7.70ms ± 0%    7.55ms ± 0%   -1.88%
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-12                    75.3ms ± 0%    67.3ms ± 0%  -10.57%
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-12                       9.92ms ± 0%    9.02ms ± 0%   -9.15%
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-12                      16.3ms ± 0%    16.1ms ± 0%   -1.46%
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-12                     83.9ms ± 0%    79.8ms ± 0%   -4.97%
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-12                     733ms ± 0%     750ms ± 0%   +2.40%
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-12                   85.5ms ± 0%    88.1ms ± 0%   +3.05%
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-12                   154ms ± 0%     148ms ± 0%   -3.84%
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-12                  793ms ± 0%     793ms ± 0%   +0.11%
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-12                 6.63s ± 0%     6.74s ± 0%   +1.62%
RangeQuery/expr=changes(a_one[1d]),steps=1-12                                       924µs ± 0%     981µs ± 0%   +6.08%
RangeQuery/expr=changes(a_one[1d]),steps=10-12                                     1.08ms ± 0%    1.14ms ± 0%   +5.71%
RangeQuery/expr=changes(a_one[1d]),steps=100-12                                    2.74ms ± 0%    3.16ms ± 0%  +15.33%
RangeQuery/expr=changes(a_one[1d]),steps=1000-12                                   19.4ms ± 0%    24.6ms ± 0%  +26.70%
RangeQuery/expr=changes(a_ten[1d]),steps=1-12                                      7.47ms ± 0%    7.34ms ± 0%   -1.76%
RangeQuery/expr=changes(a_ten[1d]),steps=10-12                                     9.30ms ± 0%    9.91ms ± 0%   +6.58%
RangeQuery/expr=changes(a_ten[1d]),steps=100-12                                    24.0ms ± 0%    30.0ms ± 0%  +24.96%
RangeQuery/expr=changes(a_ten[1d]),steps=1000-12                                    187ms ± 0%     246ms ± 0%  +31.49%
RangeQuery/expr=changes(a_hundred[1d]),steps=1-12                                  74.5ms ± 0%    84.9ms ± 0%  +13.84%
RangeQuery/expr=changes(a_hundred[1d]),steps=10-12                                 94.9ms ± 0%   107.2ms ± 0%  +13.00%
RangeQuery/expr=changes(a_hundred[1d]),steps=100-12                                 243ms ± 0%     381ms ± 0%  +57.13%
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-12                                1.87s ± 0%     3.01s ± 0%  +61.34%
RangeQuery/expr=rate(a_one[1d]),steps=1-12                                          955µs ± 0%    1147µs ± 0%  +20.11%
RangeQuery/expr=rate(a_one[1d]),steps=10-12                                        1.08ms ± 0%    1.21ms ± 0%  +11.89%
RangeQuery/expr=rate(a_one[1d]),steps=100-12                                       2.53ms ± 0%    2.51ms ± 0%   -0.92%
RangeQuery/expr=rate(a_one[1d]),steps=1000-12                                      15.3ms ± 0%    15.5ms ± 0%   +1.29%
RangeQuery/expr=rate(a_ten[1d]),steps=1-12                                         7.37ms ± 0%    9.19ms ± 0%  +24.71%
RangeQuery/expr=rate(a_ten[1d]),steps=10-12                                        8.53ms ± 0%   10.63ms ± 0%  +24.61%
RangeQuery/expr=rate(a_ten[1d]),steps=100-12                                       22.0ms ± 0%    22.8ms ± 0%   +3.63%
RangeQuery/expr=rate(a_ten[1d]),steps=1000-12                                       155ms ± 0%     146ms ± 0%   -5.26%
RangeQuery/expr=rate(a_hundred[1d]),steps=1-12                                     72.1ms ± 0%    68.8ms ± 0%   -4.57%
RangeQuery/expr=rate(a_hundred[1d]),steps=10-12                                    83.8ms ± 0%    93.2ms ± 0%  +11.27%
RangeQuery/expr=rate(a_hundred[1d]),steps=100-12                                    215ms ± 0%     222ms ± 0%   +3.12%
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-12                                   1.57s ± 0%     1.52s ± 0%   -3.42%
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1-12                              834µs ± 0%    1015µs ± 0%  +21.60%
RangeQuery/expr=absent_over_time(a_one[1d]),steps=10-12                             879µs ± 0%    1137µs ± 0%  +29.38%
RangeQuery/expr=absent_over_time(a_one[1d]),steps=100-12                           1.10ms ± 0%    1.28ms ± 0%  +15.79%
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-12                          3.32ms ± 0%    3.94ms ± 0%  +18.80%
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-12                             6.57ms ± 0%    7.21ms ± 0%   +9.74%
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-12                            7.13ms ± 0%    7.99ms ± 0%  +12.02%
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-12                           8.65ms ± 0%   10.37ms ± 0%  +19.80%
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-12                          33.8ms ± 0%    35.0ms ± 0%   +3.52%
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-12                         67.7ms ± 0%    63.2ms ± 0%   -6.66%
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-12                        69.9ms ± 0%    69.9ms ± 0%   -0.01%
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-12                       97.4ms ± 0%    94.2ms ± 0%   -3.23%
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-12                       324ms ± 0%     384ms ± 0%  +18.45%
RangeQuery/expr=-a_one,steps=1-12                                                  18.5µs ± 0%    19.9µs ± 0%   +7.71%
RangeQuery/expr=-a_one,steps=10-12                                                 17.5µs ± 0%    19.2µs ± 0%   +9.66%
RangeQuery/expr=-a_one,steps=100-12                                                30.6µs ± 0%    26.8µs ± 0%  -12.40%
RangeQuery/expr=-a_one,steps=1000-12                                                111µs ± 0%      93µs ± 0%  -15.74%
RangeQuery/expr=-a_ten,steps=1-12                                                  92.2µs ± 0%    74.6µs ± 0%  -19.04%
RangeQuery/expr=-a_ten,steps=10-12                                                 94.2µs ± 0%    91.0µs ± 0%   -3.33%
RangeQuery/expr=-a_ten,steps=100-12                                                 186µs ± 0%     176µs ± 0%   -5.19%
RangeQuery/expr=-a_ten,steps=1000-12                                                829µs ± 0%     861µs ± 0%   +3.86%
RangeQuery/expr=-a_hundred,steps=1-12                                               670µs ± 0%     663µs ± 0%   -1.06%
RangeQuery/expr=-a_hundred,steps=10-12                                              714µs ± 0%     736µs ± 0%   +3.13%
RangeQuery/expr=-a_hundred,steps=100-12                                            1.44ms ± 0%    1.86ms ± 0%  +29.07%
RangeQuery/expr=-a_hundred,steps=1000-12                                           7.87ms ± 0%    9.61ms ± 0%  +22.08%
RangeQuery/expr=a_one_-_b_one,steps=1-12                                           30.0µs ± 0%    33.2µs ± 0%  +10.60%
RangeQuery/expr=a_one_-_b_one,steps=10-12                                          36.3µs ± 0%    43.1µs ± 0%  +18.59%
RangeQuery/expr=a_one_-_b_one,steps=100-12                                          118µs ± 0%     154µs ± 0%  +29.80%
RangeQuery/expr=a_one_-_b_one,steps=1000-12                                         864µs ± 0%    1072µs ± 0%  +24.10%
RangeQuery/expr=a_ten_-_b_ten,steps=1-12                                            159µs ± 0%     161µs ± 0%   +1.46%
RangeQuery/expr=a_ten_-_b_ten,steps=10-12                                           226µs ± 0%     254µs ± 0%  +12.52%
RangeQuery/expr=a_ten_-_b_ten,steps=100-12                                          965µs ± 0%    1093µs ± 0%  +13.29%
RangeQuery/expr=a_ten_-_b_ten,steps=1000-12                                        8.18ms ± 0%   10.44ms ± 0%  +27.64%
RangeQuery/expr=a_hundred_-_b_hundred,steps=1-12                                   1.46ms ± 0%    2.18ms ± 0%  +49.17%
RangeQuery/expr=a_hundred_-_b_hundred,steps=10-12                                  2.19ms ± 0%    3.09ms ± 0%  +41.21%
RangeQuery/expr=a_hundred_-_b_hundred,steps=100-12                                 10.4ms ± 0%    13.8ms ± 0%  +32.78%
RangeQuery/expr=a_hundred_-_b_hundred,steps=1000-12                                89.5ms ± 0%   106.3ms ± 0%  +18.81%
RangeQuery/expr=a_one_-_b_one,steps=10000-12                                       8.22ms ± 0%   12.22ms ± 0%  +48.63%
RangeQuery/expr=a_ten_-_b_ten,steps=10000-12                                       80.3ms ± 0%   120.4ms ± 0%  +49.90%
RangeQuery/expr=a_hundred_-_b_hundred,steps=10000-12                                890ms ± 0%    1328ms ± 0%  +49.19%
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-12                          61.3µs ± 0%    83.2µs ± 0%  +35.84%
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-12                         64.4µs ± 0%    83.6µs ± 0%  +29.88%
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-12                         102µs ± 0%     118µs ± 0%  +15.91%
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-12                        436µs ± 0%     500µs ± 0%  +14.83%
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-12                           169µs ± 0%     202µs ± 0%  +19.29%
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-12                          197µs ± 0%     265µs ± 0%  +34.82%
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-12                         582µs ± 0%     669µs ± 0%  +15.04%
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-12                       4.09ms ± 0%    5.37ms ± 0%  +31.35%
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-12                  1.31ms ± 0%    1.28ms ± 0%   -2.54%
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-12                 1.81ms ± 0%    1.75ms ± 0%   -3.12%
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-12                7.28ms ± 0%    7.16ms ± 0%   -1.62%
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-12               53.7ms ± 0%    58.4ms ± 0%   +8.78%
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-12                           81.6µs ± 0%    73.1µs ± 0%  -10.52%
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-12                          82.2µs ± 0%    83.3µs ± 0%   +1.31%
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-12                          135µs ± 0%     141µs ± 0%   +4.71%
…